week04-TrainCNN/week-04-Pytorch1/requirements.txt
torch
Pillow
torchvision
tqdm